Changed the code to include all files in train-validation set
Here we use CNN and number of Conv layers there are 4, number of max-pool layers are 3, dropout after the final maxpool layer =0.25;
It uses 512 hidden nodes in the only hidden layer of the fully connected NN part. It uses "adam" as the optimizer and the usual loss function and activation function for image data("relu").